# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a manpower management system API.",
        "title": "Manpower API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/user/requests/{id}/decide": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ผู้อนุมัติตาม lane ของตัวเอง ทำการตัดสินใจในคำขอ (APPROVE / REJECT / RETURN)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Approvals"
                ],
                "summary": "Approve/Reject/Return a request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "request_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "action: APPROVE | REJECT | RETURN",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DecideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DecideResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.DecideRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "description": "APPROVE / REJECT / RETURN",
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "handlers.DecideResponse": {
            "type": "object",
            "properties": {
                "doc_number": {
                    "type": "string",
                    "example": "PQ24110026"
                },
                "hr_status": {
                    "type": "string",
                    "example": "WAITING_RECRUITER"
                },
                "message": {
                    "type": "string",
                    "example": "decision stored"
                },
                "origin_status": {
                    "type": "string",
                    "example": "DIR_APPROVED"
                },
                "overall_status": {
                    "type": "string",
                    "example": "IN_PROGRESS"
                },
                "request_id": {
                    "type": "integer",
                    "example": 15
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}